feat: useCartStore 영속화 기능 추가 및 성능 최적화
주요 변경사항:
- simpleCartStore를 useCartStore로 파일명 변경
- localStorage 자동 저장/복원 영속화 기능 추가
- 불필요한 useCallback/useMemo 최적화 제거 (실제 성능 문제 없음)
- 모든 컴포넌트 import 경로 정리
- 테스트에서 localStorage 초기화 로직 추가
- 100% 테스트 호환성 유지

개선된 기능:
- 페이지 새로고침 후에도 장바구니 상태 유지
- 타이머 메모리 누수 방지
- 깔끔한 코드 구조 (불필요한 최적화 제거)
