+++ src/advanced/components/Cart/Cart.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useCartStore } from '../../stores/simpleCartStore';
+import { useCartStore } from '../../stores';
 import { CartItem } from './CartItem';
 import { EmptyCart } from './EmptyCart';
