+++ src/advanced/components/Cart/CartItem.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useCartStore } from '../../stores/simpleCartStore';
+import { useCartStore } from '../../stores';
 
 interface CartItemType {
   productId: string;
