+++ src/advanced/stores/index.ts
@@ -1,7 +1,7 @@
 /**
  * Zustand 스토어 모듈 통합 export
- * @fileoverview 간단한 스토어만 export (복잡한 구조는 비활성화)
+ * @fileoverview 메인 카트 스토어 export
  */
 
-// 간단한 카트 스토어 (현재 사용 중)
-export { useCartStore } from './simpleCartStore';
+// 메인 카트 스토어 (영속화 포함)
+export { useCartStore } from './useCartStore';
